# Speed up half <-> float conversions