--- a/src/doc/oiiointro.tex
+++ b/src/doc/oiiointro.tex
@@ -237,6 +237,8 @@ Dominik Reichl \\ {\cf http://www.dominik-reichl.de/}
 {\cf http://burtleburtle.net/bob/c/lookup3.c}
 \item {\cf xxhash} \copyright 2012 Yann Collet (BSD Licensed). 
 {\cf http://code.google.com/p/xxhash/}
+\item {\cf Half} \copyright 2012-2013 Christian Rau (MIT Licensed). 
+{\cf http://sourceforge.net/projects/half/}
 \end{itemize}
 
 These other packages are all distributed under licenses that allow them