--- a/src/iv/ivimage.cpp
+++ b/src/iv/ivimage.cpp
@@ -33,7 +33,7 @@
 
 #include <boost/foreach.hpp>
 #include <boost/scoped_ptr.hpp>
-#include <half.h>
+#include <OpenEXR/half.h>
 
 #include "imageviewer.h"
 #include "strutil.h"