@lgritz 2015-04-10
Closing this as obsolete. SSE is a much better way to go.